refactor: move logic out of component
